+++ inc/usb/usbh.h
@@ -246,7 +246,7 @@ void usbh_delete_pipe(PDEVINFO, uint8_t);
 
 int32_t usbh_control_transfer(PDEVINFO, uint8_t, uint8_t, uint8_t, uint16_t, uint16_t, uint16_t, void *);
 int8_t usbh_bulk_transfer_async(PDEVINFO, uint8_t, void *, uint32_t);
-int32_t usbh_transfer_request(PURB urb);
+int8_t usbh_transfer_request(PURB urb);
 void usbh_transfer_cancel(PURB urb);
 int32_t usbh_set_configuration(PDEVINFO, uint16_t);
 int32_t usbhost_init(void);
